+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./tui/Select.js";
+export * from "./tui/MultiSelect.js";
 export * from "./filesystem.js";
 export * from "./formatting.js";
 export * from "./logger.js";